openai streaming initial refactor
